rename package of physical operators
