+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/operators/physical/CartesianProduct.java
@@ -15,36 +15,28 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.single.cypher.physical_operators;
+package org.gradoop.flink.model.impl.operators.matching.single.cypher.operators.physical;
 
 import org.apache.flink.api.java.DataSet;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
-import org.s1ck.gdl.model.cnf.CNF;
 
 /**
- * This operator joins two possibly disjunct data sets by predicates only concerning properties
- * e.g.
- * <pre>
- *   MATCH (a:Department), (b)-[:X]->(c:Person {name: "Max") WHERE a.prop = b.prop
- * </pre>
+ * Creates the Cartesian Product of two embeddings.
  */
-public class ValueJoin implements PhysicalOperator {
+public class CartesianProduct implements PhysicalOperator{
 
   private final DataSet<Embedding> lhs;
   private final DataSet<Embedding> rhs;
-  private final CNF joinCriteria;
 
   /**
-   * New value join operator
+   * New Cartesian Product Operator
    *
-   * @param lhs left hand side data set
-   * @param rhs right hand side data set
-   * @param joinCriteria join criteria
+   * @param lhs the left hand side embedding
+   * @param rhs the right hand side embedding
    */
-  public ValueJoin(DataSet<Embedding> lhs, DataSet<Embedding> rhs, CNF joinCriteria) {
+  public CartesianProduct(DataSet<Embedding> lhs, DataSet<Embedding> rhs) {
     this.lhs = lhs;
     this.rhs = rhs;
-    this.joinCriteria = joinCriteria;
   }
 
   @Override
