+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/operators/physical/FilterAndProjectVertices.java
@@ -15,40 +15,40 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.single.cypher.physical_operators;
+package org.gradoop.flink.model.impl.operators.matching.single.cypher.operators.physical;
 
 import org.apache.flink.api.java.DataSet;
-import org.gradoop.common.model.impl.pojo.Edge;
+import org.gradoop.common.model.impl.pojo.Vertex;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
 import org.s1ck.gdl.model.cnf.CNF;
 
+import java.util.List;
+
 /**
- * Filters a List of Edges by predicates and projects the remaining edges to the specified properties
- * Returns Embedding with three columns Id(sourceID), Id(Edge), Id(targetId)
+ * Filters a List of Vertices by predicates and projects the remaining vertices to the specified properties
+ * Returns one {@link org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Projection} entry
  */
-public class FilterEdges implements PhysicalOperator {
-  /**
-   * Input Edges
-   */
-  private final DataSet<Edge> input;
-  /**
-   * Predicates in conjunctive normal form
-   */
+public class FilterAndProjectVertices implements PhysicalOperator{
+
+  private final DataSet<Vertex> input;
   private final CNF predicates;
-
+  private final List<String> propertyKeys;
 
   /**
-   * New edge filter operator
-   * @param input Candidate edges
-   * @param predicates Predicates used to filter edges
+   * New Operator
+   *
+   * @param input Candidate vertices
+   * @param predicates Predicates that will be used to filter candidate vertices
+   * @param propertyKeys List of property keys that will be used for projection
    */
-  public FilterEdges(DataSet<Edge> input, CNF predicates) {
+  public FilterAndProjectVertices(DataSet<Vertex> input, CNF predicates, List<String> propertyKeys) {
     this.input = input;
     this.predicates = predicates;
+    this.propertyKeys = propertyKeys;
   }
 
+  @Override
   public DataSet<Embedding> evaluate() {
-
     return null;
   }
 }
