+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/operators/physical/FilterEdges.java
@@ -15,35 +15,40 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.single.cypher.physical_operators;
+package org.gradoop.flink.model.impl.operators.matching.single.cypher.operators.physical;
 
 import org.apache.flink.api.java.DataSet;
-import org.gradoop.common.model.impl.pojo.Vertex;
+import org.gradoop.common.model.impl.pojo.Edge;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
-
-import java.util.List;
+import org.s1ck.gdl.model.cnf.CNF;
 
 /**
- * Projects a set of vertices
- * The returned embedding consists of one entry Projection(vertex)
+ * Filters a List of Edges by predicates and projects the remaining edges to the specified properties
+ * Returns Embedding with three columns Id(sourceID), Id(Edge), Id(targetId)
  */
-public class ProjectVertices implements PhysicalOperator {
+public class FilterEdges implements PhysicalOperator {
+  /**
+   * Input Edges
+   */
+  private final DataSet<Edge> input;
+  /**
+   * Predicates in conjunctive normal form
+   */
+  private final CNF predicates;
 
-  private final DataSet<Vertex> input;
-  private final List<String> propertyKeys;
 
   /**
-   * New vertex projection operator
-   *
-   * @param input List of Vertices
-   * @param propertyKeys List of property keys that will be included in the projection
+   * New edge filter operator
+   * @param input Candidate edges
+   * @param predicates Predicates used to filter edges
    */
-  public ProjectVertices(DataSet<Vertex> input, List<String> propertyKeys) {
+  public FilterEdges(DataSet<Edge> input, CNF predicates) {
     this.input = input;
-    this.propertyKeys = propertyKeys;
+    this.predicates = predicates;
   }
 
   public DataSet<Embedding> evaluate() {
+
     return null;
   }
 }
