+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/operators/physical/ProjectVertices.java
@@ -15,39 +15,34 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.single.cypher.physical_operators;
+package org.gradoop.flink.model.impl.operators.matching.single.cypher.operators.physical;
 
 import org.apache.flink.api.java.DataSet;
-import org.gradoop.common.model.impl.pojo.Edge;
+import org.gradoop.common.model.impl.pojo.Vertex;
 import org.gradoop.flink.model.impl.operators.matching.single.cypher.embeddings.Embedding;
-import org.s1ck.gdl.model.cnf.CNF;
 
 import java.util.List;
 
 /**
- * Filters a List of Edges by predicates and projects the remaining edges to the specified properties
- * Returns Embedding with three columns Id(sourceID), Projection(Edge), Id(targetId)
+ * Projects a set of vertices
+ * The returned embedding consists of one entry Projection(vertex)
  */
-public class FilterAndProjectEdges implements PhysicalOperator{
+public class ProjectVertices implements PhysicalOperator {
 
-  private final DataSet<Edge> input;
-  private final CNF predicates;
+  private final DataSet<Vertex> input;
   private final List<String> propertyKeys;
 
   /**
-   * New Operator
+   * New vertex projection operator
    *
-   * @param input Candidate edges
-   * @param predicates Predicates that will be used to filter candidate edges
-   * @param propertyKeys List of property keys that will be used for projection
+   * @param input List of Vertices
+   * @param propertyKeys List of property keys that will be included in the projection
    */
-  public FilterAndProjectEdges(DataSet<Edge> input, CNF predicates, List<String> propertyKeys) {
+  public ProjectVertices(DataSet<Vertex> input, List<String> propertyKeys) {
     this.input = input;
-    this.predicates = predicates;
     this.propertyKeys = propertyKeys;
   }
 
-  @Override
   public DataSet<Embedding> evaluate() {
     return null;
   }
